solrClient moved to lib/ and util file created
